+++ src/connectors/configDb/configDb.connector.js
@@ -4,29 +4,68 @@ const { CONFIGDB_URL } = require("../../config");
 const { logEmitter } = require("../../services/logging.service");
 const { statusEmitter } = require("../../services/statusEmitter.service");
 
-let client;
-let configDB;
-let lcConfigCollection, deletedIdsCollection;
+let client = undefined;
+let configDB = undefined;
 
 let allLcConfigData = [];
 
-const establishConnectionToMongo = async () => {
+const establishConnectionToMongo = async collectionName => {
   if (process.env.DOUBLE_MODE === "true") {
     logEmitter.emit(
       "doubleMode",
       "configDb.connector",
       "getAllLocalCouncilConfig"
     );
-    lcConfigCollection = lcConfigCollectionDouble;
+    return lcConfigCollectionDouble;
   } else {
-    client = await mongodb.MongoClient.connect(CONFIGDB_URL, {
-      useNewUrlParser: true
+    if (configDB === undefined) {
+      client = await mongodb.MongoClient.connect(CONFIGDB_URL, {
+        useNewUrlParser: true
+      });
+      configDB = client.db("register_a_food_business_config");
+    }
+    return configDB.collection(collectionName);
+  }
+};
+
+const getConfigVersion = async regDataVersion => {
+  logEmitter.emit("functionCall", "configDb.connector", "getConfigVersion");
+
+  try {
+    const configVersionCollection = await establishConnectionToMongo(
+      "configVersion"
+    );
+    const configVersionData = await configVersionCollection.findOne({
+      _id: regDataVersion
     });
+    statusEmitter.emit("incrementCount", "getConfigFromDbSucceeded");
+    statusEmitter.emit("setStatus", "mostRecentGetConfigFromDbSucceeded", true);
+    logEmitter.emit(
+      "functionSuccess",
+      "configDB.connector",
+      "getConfigVersion"
+    );
+    return configVersionData;
+  } catch (err) {
+    statusEmitter.emit("incrementCount", "getConfigFromDbFailed");
+    statusEmitter.emit(
+      "setStatus",
+      "mostRecentGetConfigFromDbSucceeded",
+      false
+    );
 
-    configDB = client.db("register_a_food_business_config");
+    const newError = new Error();
+    newError.name = "mongoConnectionError";
+    newError.message = err.message;
 
-    lcConfigCollection = configDB.collection("lcConfig");
-    deletedIdsCollection = configDB.collection("deletedIds");
+    logEmitter.emit(
+      "functionFail",
+      "configDb.connector",
+      "getConfigVersion",
+      newError
+    );
+
+    throw newError;
   }
 };
 
@@ -39,8 +78,7 @@ const getAllLocalCouncilConfig = async () => {
 
   if (allLcConfigData.length === 0) {
     try {
-      await establishConnectionToMongo();
-
+      const lcConfigCollection = await establishConnectionToMongo("lcConfig");
       const allLcConfigDataCursor = await lcConfigCollection.find({});
       allLcConfigData = allLcConfigDataCursor.toArray();
 
@@ -86,10 +124,21 @@ const clearLcConfigCache = () => {
   return allLcConfigData;
 };
 
+const clearMongoConnection = () => {
+  client = undefined;
+  configDB = undefined;
+};
+
 const addDeletedId = async id => {
-  await establishConnectionToMongo();
+  const deletedIdsCollection = await establishConnectionToMongo("deletedIds");
 
   return deletedIdsCollection.insertOne({ id: id });
 };
 
-module.exports = { getAllLocalCouncilConfig, clearLcConfigCache, addDeletedId };
+module.exports = {
+  getAllLocalCouncilConfig,
+  clearLcConfigCache,
+  clearMongoConnection,
+  addDeletedId,
+  getConfigVersion
+};