+++ src/connectors/notify/notify.connector.js
@@ -24,11 +24,40 @@ const sendSingleEmail = async (
     flattenedData.link_to_document = pdfFile
       ? notifyClient.prepareUpload(pdfFile)
       : "";
+    const notifyTemplate = await notifyClient.getTemplateById(templateId);
+
+    const requiredTemplateFields = Object.keys(
+      notifyTemplate.body.personalisation
+    );
+
+    const templateFieldsWithoutSuffix = requiredTemplateFields.map(
+      fieldName => {
+        const trimmedFieldName = fieldName.trim();
+        return trimmedFieldName.endsWith("_exists")
+          ? trimmedFieldName.slice(0, -7)
+          : trimmedFieldName;
+      }
+    );
+
+    const templateFieldsWithoutDuplicates = new Set(
+      templateFieldsWithoutSuffix
+    );
+
+    const allNotifyPersonalisationData = { ...flattenedData };
+
+    templateFieldsWithoutDuplicates.forEach(fieldName => {
+      if (allNotifyPersonalisationData[fieldName]) {
+        allNotifyPersonalisationData[`${fieldName}_exists`] = "yes";
+      } else {
+        allNotifyPersonalisationData[fieldName] = "";
+        allNotifyPersonalisationData[`${fieldName}_exists`] = "no";
+      }
+    });
 
     const notifyArguments = [
       templateId,
       recipientEmail,
-      { personalisation: flattenedData }
+      { personalisation: allNotifyPersonalisationData }
     ];
 
     const notifyResponse = await notifyClient.sendEmail(...notifyArguments);