+++ src/connectors/notify/notify.double.js
@@ -37,7 +37,14 @@ const notifyClientDouble = {
       }
     });
   },
-  prepareUpload: () => "www.link-to-document.com"
+  prepareUpload: () => "www.link-to-document.com",
+  getTemplateById: () => ({
+    body: {
+      personalisation: {
+        establishment_trading_name: {}
+      }
+    }
+  })
 };
 
 module.exports = { notifyClientDouble };