+++ src/db/models/Establishment.js
@@ -8,7 +8,10 @@ module.exports = (sequelize, DataTypes) => {
         autoIncrement: true
       },
       establishment_trading_name: { type: DataTypes.STRING },
-      establishment_opening_date: { type: DataTypes.STRING }
+      establishment_opening_date: { type: DataTypes.STRING },
+      establishment_primary_number: { type: DataTypes.STRING },
+      establishment_secondary_number: { type: DataTypes.STRING },
+      establishment_email: { type: DataTypes.STRING }
     },
     {}
   );